requirements.txt
asgiref==3.4.1
Django==3.2.9
django-cleanup==5.2.0
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-environ==0.8.1
django-phonenumber-field==6.0.0
djangorestframework==3.12.4
phonenumbers==8.12.37
Pillow==8.4.0
psycopg2==2.9.2
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
